Add DomainController to vscene
